Move selected sensor devices to plugins

+++ drivers/QWIIC_RELAY/QWIIC_RELAY.cpp
@@ -37,7 +37,7 @@
 
 // Commands to request the state of the relay, whether it is currently on or
 // off.
- 
+
 #define RELAY_STATUS_ONE  0x05
 #define RELAY_STATUS_TWO  0x06
 #define RELAY_STATUS_THREE 0x07
@@ -60,7 +60,7 @@ QWIIC_RELAY::QWIIC_RELAY(){
 
 QWIIC_RELAY::~QWIIC_RELAY(){
     stop();
-    
+
 }
 
 
@@ -69,16 +69,16 @@ bool QWIIC_RELAY::begin(qwr_model model, bool alternateAddr){
 
     return begin(model, alternateAddr, error);
 }
- 
+
 
 bool QWIIC_RELAY::begin(qwr_model model, bool alternateAddr,  int &error){
- 
+
     uint8_t devAddr = 0;
-    
+
     _noRelays = 0;
     _model  = QWR_UNKNOWN;
-    for(int i = 0; i < sizeof(_relayState); i++) _relayState[i] = false;
-    
+    for(size_t i = 0; i < sizeof(_relayState); i++) _relayState[i] = false;
+
     switch (model) {
         case QWR_15093:  //  COM-15093  single relay
             devAddr = alternateAddr?SINGLE_ALTERNATE_ADDRESS:SINGLE_DEFAULT_ADDRESS;
@@ -98,37 +98,37 @@ bool QWIIC_RELAY::begin(qwr_model model, bool alternateAddr,  int &error){
         default:
             break;
     }
-    
+
     if(devAddr == 0) {
         LOGT_ERROR("QWIIC_RELAY model number %d unsupported", model);
         errno = EINVAL;
         error = EINVAL; // Invalid Argument.
-        
+
         return false;
     }
-    
+
     if(!_i2cPort.begin(devAddr, error))
         return  false;
- 
+
     if(!_i2cPort.smbQuick()){
         error = ENXIO;
         errno = EINVAL;
         return  false;
     }
-    
+
     _model = model;
     _isSetup = true;
-    
+
     return _isSetup;
 }
- 
+
 void QWIIC_RELAY::stop(){
 //    LOGT_INFO("QWIIC_RELAY(%02x) stop\n",  _i2cPort.getDevAddr());
 
     _isSetup = false;
     _i2cPort.stop();
 }
- 
+
 bool QWIIC_RELAY::isOpen(){
 return _i2cPort.isAvailable() && _isSetup;
 };
@@ -139,55 +139,55 @@ uint8_t    QWIIC_RELAY::getDevAddr(){
     return _i2cPort.getDevAddr();
 };
 
- 
+
 bool QWIIC_RELAY::allOn(){
     bool status = false;
-    
+
     if(_model == QWR_15093){
         status = setRelay(1, true);
     }
     else {
         uint8_t cmd[1] = {TURN_ALL_ON};
-        
+
         status = _i2cPort.stdWriteBytes(sizeof(cmd),cmd);
-        
+
         updateStatus();
     }
-    
+
     if(!status){
         LOGT_ERROR("QWIIC_RELAY(%02X) ALL ON failed: %s",
                    _i2cPort.getDevAddr(),  strerror(errno));
     }
-    
+
     return status;
 };
 
 
 bool QWIIC_RELAY::allOff(){
     bool status = false;
- 
+
     if(_model == QWR_15093){
         status = setRelay(1, false);
     }
     else {
         uint8_t cmd[1] = {TURN_ALL_OFF};
-        
+
         status = _i2cPort.stdWriteBytes(sizeof(cmd),cmd);
-        
+
         updateStatus();
     }
-    
+
     if(!status){
         LOGT_ERROR("QWIIC_RELAY(%02X) ALL OFF failed: %s",
                    _i2cPort.getDevAddr(),  strerror(errno));
     }
-    
+
     return status;
 };
 
 
 bool QWIIC_RELAY::relayOn(uint8_t relayNum){
- 
+
    return setRelay(relayNum, true);
  };
 
@@ -197,37 +197,37 @@ bool QWIIC_RELAY::relayOff(uint8_t relayNum){
 
 
 bool  QWIIC_RELAY::setRelay(uint8_t relayNum, bool state){
-  
+
     bool status = false;
-    
+
     if(relayNum == 0 || relayNum > _noRelays) {
         LOGT_ERROR("QWIIC_RELAY(%02X) relayOn(%d) param error",
                    _i2cPort.getDevAddr(), relayNum);
         return false;
     }
-  
+
     if(_model == QWR_15093){
-        
+
          #define COMMAND_RELAY_OFF             0x00
          #define COMMAND_RELAY_ON               0x01
 
         uint8_t cmd[1] = {0};
         cmd[0] = state?COMMAND_RELAY_ON:COMMAND_RELAY_OFF;
-        
+
         status = _i2cPort.stdWriteBytes(sizeof(cmd),cmd);
 
     }
     else
     {
         updateStatus();
-  
+
         if( _relayState[relayNum -1] == state )
             return true;
-     
+
         /*
-         
+
          relayNum is the same as the command
-         
+
          #define RELAY_ONE_TOGGLE 0x01
          #define RELAY_TWO_TOGGLE 0x02
          #define RELAY_THREE_TOGGLE 0x03
@@ -235,14 +235,14 @@ bool  QWIIC_RELAY::setRelay(uint8_t relayNum, bool state){
          */
         uint8_t cmd[1] = {relayNum};
         status = _i2cPort.stdWriteBytes(sizeof(cmd),cmd);
-      
+
     }
-   
+
     if(!status){
         LOGT_ERROR("QWIIC_RELAY(%02X) WRITE failed: %s",
                    _i2cPort.getDevAddr(),  strerror(errno));
     }
- 
+
     return status;
 }
 
@@ -250,35 +250,35 @@ bool  QWIIC_RELAY::setRelay(uint8_t relayNum, bool state){
 
 bool QWIIC_RELAY::relayState(std::vector<bool> &state){
     bool status = false;
-    
+
     status = updateStatus();
-  
+
     if(status){
         state.clear();
-        
+
         for(int i = 0; i < _noRelays; i++)
             state.push_back(_relayState[i]);
     }
-    
+
     return status;
 }
 
 bool QWIIC_RELAY::updateStatus(){
-    
+
     bool status = false;
-    
+
 #if 0
     I2C::i2c_block_t block;
     status = _i2cPort.readBlock(RELAY_STATUS_ONE, _noRelays, block);
-    
+
     if(status){
         for(int i = 0; i < _noRelays; i++){
             _relayState[i] = block[i]?true:false;
         }
         return true;
     }
-    
-    
+
+
 #else
     uint8_t stat[4] = {0};
     status = _i2cPort.readBytes(RELAY_STATUS_ONE, _noRelays ,stat);
@@ -286,10 +286,10 @@ bool QWIIC_RELAY::updateStatus(){
         for(int i = 0; i < _noRelays; i++){
             _relayState[i] = stat[i]?true:false;
         }
-        
+
         return true;
     }
 #endif
-    
+
     return false;
 }
